requirements.txt
pandas
scikit-learn
spacy
sentence-transformers
datasets
pillow
numpy
